first bit of the RPC dispatcher converted to using commands and/or facades
